added detection of last-chunks for chunked transfer
RFC2616, section 3.6.1 [1] states, that the
last-chunk has the chunk-size "0" (and possible
a chunk-extension, which MUST be ignored, if not
understood - hence the regexp) terminated by \r\n.
After the last-chunk there can be an OPTIONAL
trailer, which does not belong to the chunked
payload.

Yajl::HTTPStream#request now respects the RFC and
stops reading after the last-chunk which led to a
timeout before.

[1] http://tools.ietf.org/html/rfc2616#section-3.6.1