Remove JSON gem compatibility API
This API causes much more trouble than it's worth. At the time of
writing, there are at least three JSON libraries for Ruby which
implement the `to_json` method. They all treat it slightly different
by assuming different parameters being passed in. Instead of try and
support all of those use-cases I've decided to drop it from yajl-ruby.

I may try and add something to the examples/ dir which you could
"drop-in" to an application to have this functionality again.