Add a string cache to take advantage of CoW strings
This patch adds an st_table to cache strings as we parse JSON.  The
first time a string is parsed, if that string is longer than the
embedded string length, it looks up the string in the string cache hash.
If the string is cached, it returns a duplicate string.  If the string
isn't cached, it allocates a new Ruby string and caches the string in
the hash.  This takes advantage of non-embedded CoW Ruby strings.  IOW,
we only memcopy the bytes once for each unique string.