Fix leak when non-object or array is passed to parser
If something besides an object or array is passed to the parser, it will
raise an exception and the lexer isn't free'd.  This patch moves the
token checks up a little so that we can free the lexer before raising an
exception.

This program will leak memory without bounds:

```ruby
require 'yajl'

loop do
  begin
    stream = StringIO.new('foo')
    projector = Yajl::Projector.new(stream)
    projector.project({"name" => nil})
  rescue Yajl::ParseError
  end
end
```