Add as_json to yajl
Support as_json so we can change Rails' json implementation and this
method is required. Yajl used to not support as_json because there was
an infinite loop, but here if the object returns from as_json is not the
one that we called as_json on then we dump that.