Add gem build and install steps to CI
I think this is a Ruby 3.0 issue...locally I had to build and install
the gem for the spec/parsing/large_number_spec.rb tests to pass. They
were failing to find `yajl` in the `require` call. I reproduced this in
the console as well. I believe this has to do with changes to bundler in
Ruby 3.0 but I'm not 100% sure.